requirements.txt
pandas>=0.23.4

pyzotero==1.4.26
pyarrow==5.0.0
tqdm==4.64.0
urllib3==1.26.12

google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib

python-dotenv